+++ packages/smath/CHANGELOG.md
@@ -5,6 +5,7 @@
 - Add fraction functions `rat()` to decompose ratios and `mixed()` to decompose mixed numbers
 - Update dependency versions
 - Remove `rseq()` support
+- Update functions in `npx` binary
 
 ## 1.10.0